+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Features:
 * Add `--docker-option` to pass raw options to Docker
 * Allow `download <name>`, shortcut for `download <name>:./<name>` (downloads to current directory, keep name)
 * Add `--input` and `--output` to showfiles
+* You can use `run -` to run every run, now even if there is only one run
 
 1.0.3 (2015-12-02)
 ------------------