+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Bugfixes:
 
 Features:
 * Add `--docker-option` to pass raw options to Docker
+* You can use `run -` to run every run, now even if there is only one run
 
 1.0.3 (2015-12-02)
 ------------------