+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Features:
 * Allow `download <name>`, shortcut for `download <name>:./<name>` (downloads to current directory, keep name)
 * Allow `download --all`
 * Add `--input` and `--output` to showfiles
-* You can use `run -` to run every run, now even if there is only one run
+* You can use `run` or `run -` to run every run, regardless of their number
 
 1.0.3 (2015-12-02)
 ------------------